project/processing-service/requirements.txt
# Requirements.txt
requests
fastapi==0.104.1
uvicorn==0.24.0
confluent-kafka==2.3.0
alpaca-trade-api==3.0.0
python-dotenv==1.0.0
pytz==2023.3
fastavro
pandas
apache-flink==1.19.0